refactor: consolidate server cookie automation code under lib/automation/cookie
Pure move — no method names, signatures, or implementation details change.

The server's cookie automation code was spread across lib/automation/cookies.ts,
lib/automation/util.ts, lib/util/cookies.ts, and inline in the three
browsers/*_automation.ts files. It now lives together under
lib/automation/cookie/:

- automation.ts — the Cookies class (the automation-facing layer that
  normalizes the automate mechanism into the cookie APIs), from
  lib/automation/cookies.ts
- util.ts — cookieMatches, isHostOnlyCookie, CyCookie, CyCookieFilter, from
  lib/automation/util.ts (+ isHostOnlyCookie from cdp_automation.ts)
- jar.ts — CookieJar and tough-cookie serialization, from lib/util/cookies.ts
- converters/{cdp,bidi,webkit}.ts — the per-protocol cookie type conversion
  helpers, out of the browsers/*_automation.ts files

Mechanical consequences only: functions that now cross a file boundary are
exported, importers (including the proxy/driver/runner deep imports of the
former lib/util/cookies) re-point to the new paths, and
test/unit/automation/util_spec.ts moves to test/unit/automation/cookie/util_spec.ts.

One type-level fix the move surfaced: proxy's service-worker-injector.ts was
compiling against a WorkerGlobalScope interface that entered its program
accidentally — from @types/filesystem via @types/chrome, referenced through
the old deep-import chain into server's cdp_automation. With that chain
severed, the file now declares a file-scoped structural WorkerGlobalScope
stand-in (matching its existing FetchEvent/ExtendableEvent pattern) instead
of relying on the ambient type.
