+++ packages/proxy/lib/http/util/cookies.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash'
 import type Debug from 'debug'
 import { URL } from 'url'
-import { SerializableAutomationCookie, Cookie, CookieJar, toughCookieToAutomationCookie } from '@packages/server/lib/util/cookies'
+import { SerializableAutomationCookie, Cookie, CookieJar, toughCookieToAutomationCookie } from '@packages/server/lib/automation/cookie/jar'
 import type { RequestCredentialLevel, ResourceType } from '../../resourceTypeAndCredentialManager'
 
 import { urlOriginsMatch, urlSameSiteMatch } from '@packages/network-tools'
