+++ packages/proxy/lib/http/util/service-worker-injector.ts
@@ -2,8 +2,13 @@
 
 import type { ServiceWorkerClientEvent } from './service-worker-manager'
 
+// structural stand-in for the webworker lib's WorkerGlobalScope — the real lib can't be
+// referenced because lib="webworker" conflicts with the dom lib program-wide
+interface WorkerGlobalScope extends EventTarget {}
+
 // this should be of type ServiceWorkerGlobalScope from the webworker lib,
-// but we can't reference it directly because it causes errors in other packages
+// but we can't reference it directly because it causes errors in other packages,
+// so the members the injected code uses are declared explicitly here
 interface ServiceWorkerGlobalScope extends WorkerGlobalScope {
   registration: ServiceWorkerRegistration
   clients: {
