+++ packages/proxy/test/unit/http/request-middleware.spec.ts
@@ -8,7 +8,7 @@ import { testMiddleware } from './helpers'
 import { CypressIncomingRequest, CypressOutgoingResponse } from '../../../lib'
 import { HttpBuffer, HttpBuffers } from '../../../lib/http/util/buffers'
 import { RemoteStates, DocumentDomainInjection } from '@packages/network-tools'
-import { CookieJar } from '@packages/server/lib/util/cookies'
+import { CookieJar } from '@packages/server/lib/automation/cookie/jar'
 import { NetworkInterceptionCore } from '@packages/network-interception'
 import { HttpMiddlewareThis } from '../../../lib/http'
 import { resourceTypeAndCredentialManager } from '../../../lib/resourceTypeAndCredentialManager'
