+++ packages/server/lib/automation/automation.ts
@@ -1,10 +1,10 @@
 import Bluebird from 'bluebird'
 import { randomUUID } from 'crypto'
-import { Cookies } from './cookies'
+import { Cookies } from './cookie/automation'
 import { Screenshot } from './screenshot'
 import type { BrowserPreRequest } from '@packages/proxy'
 import type { AutomationCommands, AutomationMiddleware, OnRequestEvent, OnServiceWorkerClientSideRegistrationUpdated, OnServiceWorkerRegistrationUpdated, OnServiceWorkerVersionUpdated } from '@packages/types'
-import { cookieJar } from '../util/cookies'
+import { cookieJar } from './cookie/jar'
 import type { ServiceWorkerEventHandler } from '@packages/proxy/lib/http/util/service-worker-manager'
 import Debug from 'debug'
 import { AutomationNotImplemented } from './automation_not_implemented'
