+++ packages/server/lib/browsers/bidi_automation.ts
@@ -1,9 +1,7 @@
 import debugModule from 'debug'
-import toInteger from 'lodash/toInteger'
-import isNumber from 'lodash/isNumber'
-import { isHostOnlyCookie } from './cdp_automation'
 import { isLocalhost } from '@packages/network-tools'
-import { cookieMatches } from '../automation/util'
+import { cookieMatches, isHostOnlyCookie } from '../automation/cookie/util'
+import { convertBiDiCookieToCyCookie, convertCyCookieToBiDiCookie, convertSameSiteExtensionToBiDi } from '../automation/cookie/converters/bidi'
 import { bidiKeyPress } from '../automation/commands/key_press'
 import { AutomationNotImplemented } from '../automation/automation_not_implemented'
 
@@ -17,11 +15,10 @@ import type {
   NetworkResponseStartedParameters,
   NetworkResponseCompletedParameters,
   NetworkFetchErrorParameters,
-  NetworkCookie,
   BrowsingContextInfo,
-  NetworkSameSite,
 } from 'webdriver/build/bidi/localTypes'
-import type { CyCookie as CyBaseCookie } from '../automation/util'
+import type { CyCookie as CyBaseCookie } from '../automation/cookie/util'
+import type { BidiCyCookie as CyCookie, StoragePartialCookie } from '../automation/cookie/converters/bidi'
 import { bidiGetUrl } from '../automation/commands/get_url'
 import { bidiReloadFrame } from '../automation/commands/reload_frame'
 import { bidiNavigateHistory } from '../automation/commands/navigate_history'
@@ -36,28 +33,6 @@ const BIDI_SCREENSHOT_DEBUG_NAMESPACE = `${BIDI_DEBUG_NAMESPACE}:screenshot`
 const debug = debugModule(BIDI_DEBUG_NAMESPACE)
 const debugCookies = debugModule(BIDI_COOKIE_DEBUG_NAMESPACE)
 const debugScreenshot = debugModule(BIDI_SCREENSHOT_DEBUG_NAMESPACE)
-
-type CyCookie = Omit<CyBaseCookie, 'sameSite'> & {
-  sameSite: 'no_restriction' | 'lax' | 'strict' | 'unspecified'
-}
-
-// if the filter is not an exact match OR, if looselyMatchCookiePath is enabled, doesn't include the path.
-// ex: /foo/bar/baz path should include cookies for /foo/bar/baz, /foo/bar, /foo, and /
-// this is shipped in remoteTypes within webdriver but it isn't exported, so we need to redefine the type
-interface StoragePartialCookie extends Record<string, unknown> {
-  name: string
-  value: {
-    type: 'string'
-    value: string
-  }
-  domain: string
-  path: string
-  httpOnly: boolean
-  hostOnly?: boolean
-  secure: boolean
-  sameSite: NetworkSameSite | 'default'
-  expiry?: number
-}
 
 const debugVerbose = debugModule('cypress-verbose:server:browsers:bidi_automation')
 
@@ -103,87 +78,6 @@ const normalizeResourceType = (type: RequestInitiatorType): ResourceType => {
     default:
       return type
   }
-}
-
-function convertSameSiteBiDiToExtension (str: NetworkSameSite | 'default') {
-  if (str === 'none') {
-    return 'no_restriction'
-  }
-
-  if (str === 'default') {
-    // put firefox version check here, under 140 we need to return 'no_restriction'
-    return 'unspecified'
-  }
-
-  return str
-}
-
-function convertSameSiteExtensionToBiDi (str: CyCookie['sameSite'], majorFirefoxVersion?: number) {
-  if (str === 'no_restriction') {
-    return 'none'
-  }
-
-  if (str === 'unspecified') {
-    // put firefox version check here, under 140 we need to return 'no_restriction'
-    return 'default'
-  }
-
-  // @see https://www.w3.org/TR/webdriver-bidi/#type-network-Cookie
-  // in Firefox 140, BiDi added the 'default' value to be able to assign 'unspecified', which was also added in Firefox 140.
-  const defaultValue = majorFirefoxVersion && majorFirefoxVersion < 140 ? 'none' : 'default'
-
-  // if no value, default to 'none' as this is the browser default in firefox specifically.
-  // Every other browser defaults to 'lax'
-  return str === undefined ? defaultValue : str
-}
-
-// used to normalize cookies to CyCookie before returning them through the automation client
-const convertBiDiCookieToCyCookie = (cookie: NetworkCookie): CyCookie => {
-  const cyCookie: CyCookie = {
-    name: cookie.name,
-    value: cookie.value.value,
-    domain: cookie.domain,
-    path: cookie.path,
-    httpOnly: cookie.httpOnly,
-    hostOnly: !!isHostOnlyCookie(cookie),
-    expirationDate: cookie.expiry ?? undefined,
-    secure: cookie.secure,
-    sameSite: convertSameSiteBiDiToExtension(cookie.sameSite),
-  }
-
-  debugCookies(`parsed BiDi cookie %o to cy cookie %o`, cookie, cyCookie)
-
-  return cyCookie
-}
-
-const convertCyCookieToBiDiCookie = (cookie: CyCookie, majorFirefoxVersion?: number): StoragePartialCookie => {
-  const cookieToSet: StoragePartialCookie = {
-    name: cookie.name,
-    value: {
-      type: 'string',
-      value: cookie.value,
-    },
-    domain: cookie.domain,
-    path: cookie.path,
-    httpOnly: cookie.httpOnly,
-    secure: cookie.secure,
-    sameSite: convertSameSiteExtensionToBiDi(cookie.sameSite, majorFirefoxVersion),
-    // BiDi cookie expiry is in seconds from EPOCH, but sometimes the automation client feeds in a float and BiDi does not know how to handle it.
-    // If trying to set a float on the expiry time in BiDi, the setting silently fails.
-    expiry: (cookie.expirationDate === -Infinity ? 0 : (isNumber(cookie.expirationDate) ? toInteger(cookie.expirationDate) : null)) ?? undefined,
-  }
-
-  if (!cookie.hostOnly && isHostOnlyCookie(cookie)) {
-    cookieToSet.domain = `.${cookie.domain}`
-  }
-
-  if (cookie.hostOnly && !isHostOnlyCookie(cookie)) {
-    cookieToSet.hostOnly = false
-  }
-
-  debugCookies(`parsed cy cookie %o to BiDi cookie %o`, cookie, cookieToSet)
-
-  return cookieToSet
 }
 
 const buildBiDiClearCookieFilterFromCyCookie = (cookie: CyCookie, majorFirefoxVersion?: number): StoragePartialCookie => {
