+++ packages/server/lib/open_project.ts
@@ -9,7 +9,7 @@ import * as errors from './errors'
 import preprocessor from './plugins/preprocessor'
 import runEvents from './plugins/run_events'
 import * as session from './session'
-import { cookieJar } from './util/cookies'
+import { cookieJar } from './automation/cookie/jar'
 import { getSpecUrl } from './project_utils'
 import type { BrowserLaunchOpts, OpenProjectLaunchOptions, InitializeProjectOptions, OpenProjectLaunchOpts, FoundBrowser, AutomationCommands } from '@packages/types'
 import { DataContext, getCtx } from '@packages/data-context'
