+++ packages/server/lib/server-base.ts
@@ -34,7 +34,7 @@ import { InitializeRoutes, createCommonRoutes } from './routes'
 import type { FoundSpec, ProtocolManagerShape, TestingType } from '@packages/types'
 import { RemoteStates } from '@packages/network-tools'
 import type { RemoteState } from '@packages/network-tools'
-import { cookieJar, SerializableAutomationCookie } from './util/cookies'
+import { cookieJar, SerializableAutomationCookie } from './automation/cookie/jar'
 import * as fileServer from './file_server'
 import type { FileServer } from './file_server'
 import * as appData from './util/app_data'
@@ -47,7 +47,7 @@ import isHtml from 'is-html'
 import type Protocol from 'devtools-protocol'
 import type { ServiceWorkerClientEvent } from '@packages/proxy/lib/http/util/service-worker-manager'
 import type { Automation } from './automation'
-import type { AutomationCookie } from './automation/cookies'
+import type { AutomationCookie } from './automation/cookie/automation'
 import type { ResourceType, RequestCredentialLevel } from '@packages/proxy'
 import { GracefulExit } from './util/graceful-exit'
 import { createProxyRuntime } from './network-runtime'
